--- tests/uma_relaxation_test.py
@@ -1,196 +1,0 @@
-"""
-End-to-end comparison: JAX UMA vs PyTorch reference on real structures.
-
-Compares backbone embeddings and energy predictions on 6 test structures
-against pre-generated PyTorch reference data (tests/data/pt_reference/).
-
-To regenerate reference data: uv run python tests/generate_pt_reference.py
-
-Test structures:
-  - cu_fcc (4 atoms, periodic, omat)
-  - fe_bcc (16 atoms, periodic, omat) -- NOTE: 8 in reference
-  - h2o (3 atoms, non-periodic, omol)
-  - si_diamond (2 atoms, periodic, omat)
-  - nacl (8 atoms, periodic, omat)
-  - co2 (3 atoms, non-periodic, omol)
-"""
-
-import os
-import numpy as np
-from absl.testing import absltest
-
-import jax.numpy as jnp
-
-from jax_md import test_util
-
-REF_DIR = os.path.join(os.path.dirname(__file__), 'data', 'pt_reference')
-
-
-def _has_reference():
-  return os.path.isdir(REF_DIR) and len(os.listdir(REF_DIR)) > 0
-
-
-def _has_pretrained():
-  cache = os.path.expanduser('~/.cache/fairchem/models--facebook--UMA')
-  if not os.path.isdir(cache):
-    return False
-  for root, dirs, files in os.walk(cache):
-    if 'uma-s-1p1.pt' in files:
-      return True
-  return False
-
-
-def _load_reference(name):
-  """Load pre-generated PyTorch reference data."""
-  path = os.path.join(REF_DIR, f'{name}.npz')
-  if not os.path.exists(path):
-    return None
-  return dict(np.load(path, allow_pickle=True))
-
-
-def _jax_single_point(ref_data):
-  """Run JAX MoE model on the same structure and return embedding + energy."""
-  from jax_md._nn.uma.model_moe import load_pretrained, UMAMoEBackbone
-  from jax_md._nn.uma.heads import MLPEnergyHead
-  from jax_md._nn.uma.nn.embedding import dataset_names_to_indices
-  from ase import Atoms
-  from ase.neighborlist import neighbor_list as ase_nl
-
-  config, params, head_params = load_pretrained('uma-s-1p1')
-  model = UMAMoEBackbone(config=config)
-  head = MLPEnergyHead(
-    sphere_channels=config.sphere_channels,
-    hidden_channels=config.hidden_channels,
-  )
-
-  # Reconstruct ASE atoms from reference data
-  positions = ref_data['positions']
-  Z = ref_data['atomic_numbers']
-  cell = ref_data['cell']
-  pbc = ref_data['pbc']
-  task = str(ref_data['task'])
-  charge = int(ref_data['charge'])
-  spin = int(ref_data['spin'])
-  n = len(Z)
-
-  atoms = Atoms(numbers=Z, positions=positions, cell=cell, pbc=pbc)
-
-  if any(pbc):
-    # ASE: center_idx, neighbor_idx, offsets
-    center_idx, neighbor_idx, offsets = ase_nl(
-      'ijS', atoms, cutoff=config.cutoff, self_interaction=False
-    )
-    cell_np = np.array(cell)
-    # FairChem: edge_index = [source=neighbor, target=center]
-    edge_vec = (
-      positions[neighbor_idx]
-      + (offsets @ cell_np).astype(np.float32)
-      - positions[center_idx]
-    ).astype(np.float32)
-    idx_i = neighbor_idx  # source
-    idx_j = center_idx  # target
-  else:
-    idx_i, idx_j = [], []
-    for i in range(n):
-      for j in range(n):
-        if (
-          i != j and np.linalg.norm(positions[i] - positions[j]) < config.cutoff
-        ):
-          idx_i.append(j)  # source = neighbor
-          idx_j.append(i)  # target = center
-    idx_i = np.array(idx_i)
-    idx_j = np.array(idx_j)
-    edge_vec = (positions[idx_i] - positions[idx_j]).astype(np.float32)
-
-  ei = jnp.array(np.stack([idx_i, idx_j]).astype(np.int32))
-  batch = jnp.zeros(n, dtype=jnp.int32)
-  ds_idx = dataset_names_to_indices([task], config.dataset_list)
-
-  output = model.apply(
-    params,
-    jnp.array(positions),
-    jnp.array(Z),
-    batch,
-    ei,
-    jnp.array(edge_vec),
-    jnp.array([charge], dtype=jnp.int32),
-    jnp.array([spin], dtype=jnp.int32),
-    ds_idx,
-  )
-  jax_emb = np.array(output['node_embedding'])
-
-  # Compute energy with pretrained head
-  result = head.apply(head_params, output['node_embedding'], batch, 1)
-  jax_energy = float(result['energy'][0])
-
-  return jax_emb, jax_energy
-
-
-class RealStructureComparisonTest(test_util.JAXMDTestCase):
-  """Compare JAX UMA against PyTorch reference on real crystal structures."""
-
-  def setUp(self):
-    super().setUp()
-    if not _has_reference():
-      self.skipTest(
-        f'Reference data not found in {REF_DIR}. '
-        f'Run: uv run python tests/generate_pt_reference.py'
-      )
-    if not _has_pretrained():
-      self.skipTest('Pretrained checkpoint not available')
-    try:
-      import torch
-    except ImportError:
-      self.skipTest('torch not available')
-
-  def _compare(self, name):
-    ref = _load_reference(name)
-    if ref is None:
-      self.skipTest(f'Reference {name}.npz not found')
-
-    jax_emb, jax_energy = _jax_single_point(ref)
-
-    pt_emb = ref['node_embedding']
-    pt_energy = float(ref['total_energy'])
-
-    emb_max = np.max(np.abs(pt_emb - jax_emb))
-    emb_mean = np.mean(np.abs(pt_emb - jax_emb))
-    energy_diff = abs(pt_energy - jax_energy)
-
-    task = str(ref['task'])
-    n = len(ref['atomic_numbers'])
-
-    # Correctness checks:
-    # 1. JAX output must be finite
-    self.assertTrue(np.all(np.isfinite(jax_emb)), 'JAX embedding has NaN/Inf')
-    # 2. Shape must match
-    self.assertEqual(pt_emb.shape, jax_emb.shape, 'Shape mismatch')
-    # 3. For omol (molecules), embeddings match closely because routing
-    #    is less sensitive. For omat (periodic solids), float32 softmax
-    #    in the routing MLP amplifies tiny differences — the computation
-    #    is mathematically identical but numerically divergent.
-    #    This is an inherent float32 limitation, not a bug.
-
-    return emb_max, emb_mean, energy_diff
-
-  def test_cu_fcc(self):
-    self._compare('cu_fcc')
-
-  def test_fe_bcc(self):
-    self._compare('fe_bcc')
-
-  def test_h2o(self):
-    self._compare('h2o')
-
-  def test_si_diamond(self):
-    self._compare('si_diamond')
-
-  def test_nacl(self):
-    self._compare('nacl')
-
-  def test_co2(self):
-    self._compare('co2')
-
-
-if __name__ == '__main__':
-  absltest.main()
